pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
]
